+++ service/company/company_server/src/main/java/com/sparta/company/application/dto/company/CompanyResponse.java
@@ -1,22 +1,16 @@
-package com.sparta.company.application.dto;
+package com.sparta.company.application.dto.company;
 
 import com.sparta.company.domain.CompanyType;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 import lombok.Data;
 
 @Data
-public class CompanyCreateRequest {
+public class CompanyResponse {
 
-  @NotNull
+  private UUID companyId;
   private Long userId;
-  @NotNull
   private UUID hubId;
-  @NotBlank
   private String companyName;
-  @NotNull
   private CompanyType companyType;
-  @NotBlank
   private String companyAddress;
 }
