+++ service/company/company_server/src/main/java/com/sparta/company/application/dto/company/CompanySearchCond.java
@@ -1,9 +1,8 @@
-package com.sparta.company.application.dto;
+package com.sparta.company.application.dto.company;
 
 import com.sparta.company.domain.CompanyType;
 import java.util.UUID;
 import lombok.Data;
-import lombok.Getter;
 
 @Data
 public class CompanySearchCond {
