+++ service/company/company_server/src/main/java/com/sparta/company/application/dto/company/CompanyUpdateRequest.java
@@ -1,0 +1,21 @@
+package com.sparta.company.application.dto.company;
+
+import com.sparta.company.domain.CompanyType;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import java.util.UUID;
+import lombok.Data;
+
+@Data
+public class CompanyUpdateRequest {
+  @NotNull
+  private Long userId;
+  @NotNull
+  private UUID hubId;
+  @NotBlank
+  private String companyName;
+  @NotBlank
+  private String companyAddress;
+  @NotNull
+  private CompanyType companyType;
+}
