+++ service/company/company_server/src/main/java/com/sparta/company/exception/GlobalExceptionHandler.java
@@ -1,7 +1,11 @@
 package com.sparta.company.exception;
 
+import com.sparta.commons.domain.exception.BusinessException;
 import com.sparta.commons.domain.response.FailedResponseBody;
+import com.sparta.commons.domain.response.ResponseBody;
+import com.sparta.company.exception.AlreadyDeletedException;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -9,6 +13,14 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+    @ExceptionHandler(BusinessException.class) // custom 에러
+    public ResponseEntity<ResponseBody<Void>> handleServiceException(HttpServletRequest request,
+        BusinessException e) {
+        com.sparta.commons.domain.exception.ErrorCode errorCode = e.getErrorCode();
+        return ResponseEntity.status(errorCode.getStatus())
+            .body(new FailedResponseBody(errorCode.getCode(), errorCode.getMessage()));
+    }
 
     @ExceptionHandler(AlreadyDeletedException.class)
     public ResponseEntity<FailedResponseBody> handleAlreadyDeletedException(AlreadyDeletedException e) {
